# docs/api/xorb-openapi-spec.yaml
openapi: 3.1.0
info:
  title: XORB Cybersecurity Platform API
  description: |
    XORB API for autonomous cybersecurity operations, orchestration, and multi-agent coordination.
    
    ## Security
    - All endpoints require mTLS authentication
    - JWT bearer tokens for authorization
    - Role-based access control (RBAC)
    - Request signing for response integrity
    
    ## Rate Limiting
    - Admin: 10,000 req/min
    - Orchestrator: 5,000 req/min  
    - Agent: 2,000 req/min
    - Analyst: 1,000 req/min
    - Readonly: 500 req/min
    
  version: 1.0.0
  contact:
    name: XORB Platform Team
    email: platform@xorb.security
  license:
    name: Proprietary
servers:
  - url: https://api.xorb.security/v1
    description: Production API
  - url: https://staging-api.xorb.security/v1
    description: Staging API
  - url: http://localhost:8000/v1
    description: Development API
security:
  - BearerAuth: []
  - mTLS: []
paths:
...
  /agents:
    get:
      summary: List all agents
      description: Retrieve list of all registered agents
      tags: [Agent Management]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by agent status
          schema:
            $ref: '#/components/schemas/AgentStatus'
        - name: type
          in: query
          description: Filter by agent type
          schema:
            type: string
        - name: limit
          in: query
          description: Limit number of results
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Agents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create new agent
      description: Deploy a new autonomous agent
      tags: [Agent Management]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mTLS:
      type: mutualTLS
      description: Client certificate authentication
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [security_analyst, red_team, blue_team, threat_hunter, compliance_monitor]
        status:
          $ref: '#/components/schemas/AgentStatus'
        capabilities:
          type: array
          items:
            type: string
        configuration:
          type: object
        created_at:
          type: string
          format: date-time
        last_active:
          type: string
          format: date-time
        metrics:
          $ref: '#/components/schemas/AgentMetrics'
    AgentStatus:
      type: string
      enum: [active, idle, busy, error, terminated]
    AgentMetrics:
      type: object
      properties:
        tasks_completed:
          type: integer
        success_rate:
          type: number
          format: float
        avg_response_time:
          type: number
          format: float
        cpu_usage:
          type: number
          format: float
        memory_usage:
          type: number
          format: float
    CreateAgentRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        capabilities:
          type: array
          items:
            type: string
        configuration:
          type: object
...
    AgentsListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...