# src/api/app/routers/telemetry.py
"""
XORB Telemetry & Monitoring API Endpoints
Provides system health, metrics, and observability endpoints
"""
import asyncio
import time
from datetime import datetime, timedelta
from typing import Dict, Any, List, Optional

from fastapi import APIRouter, Depends, Query
from pydantic import BaseModel, Field

from ..security import (
    SecurityContext,
    get_security_context,
    require_permission,
    Permission
)


# Pydantic Models
class ComponentHealth(BaseModel):
    """Individual component health status"""
    status: str  # healthy, degraded, unhealthy
    details: str
    last_check: datetime
    uptime_seconds: Optional[int] = None
    error_count: int = 0


class HealthResponse(BaseModel):
    """System health response"""
    status: str  # healthy, degraded, unhealthy
    components: Dict[str, ComponentHealth]
    timestamp: datetime
    overall_uptime_seconds: int
    version: str = "1.0.0"


class Metric(BaseModel):
    """Individual metric data point"""
    name: str
    value: float
    unit: str
    tags: Dict[str, str] = Field(default_factory=dict)
    timestamp: datetime


class MetricsResponse(BaseModel):
    """System metrics response"""
    metrics: List[Metric]
    timestamp: datetime
    collection_interval_seconds: int = 60


# System state tracking
system_start_time = time.time()
component_states = {
    "api_gateway": {
        "status": "healthy",
        "last_check": datetime.utcnow(),
        "uptime": 0,
        "errors": 0
    },
    "database": {
        "status": "healthy", 
        "last_check": datetime.utcnow(),
        "uptime": 0,
        "errors": 0
    },
    "redis_cache": {
        "status": "healthy",
        "last_check": datetime.utcnow(), 
        "uptime": 0,
        "errors": 0
    },
    "agent_manager": {
        "status": "healthy",
        "last_check": datetime.utcnow(),
        "uptime": 0,
        "errors": 0
    },
    "orchestrator": {
        "status": "healthy",
        "last_check": datetime.utcnow(),
        "uptime": 0,
        "errors": 0
    },
    "security_engine": {
        "status": "healthy",
        "last_check": datetime.utcnow(),
        "uptime": 0,
        "errors": 0
    },
    "intelligence_brain": {
        "status": "healthy",
        "last_check": datetime.utcnow(),
        "uptime": 0,
        "errors": 0
    }
}


router = APIRouter(prefix="/telemetry", tags=["Telemetry & Monitoring"])


@router.get("/health", response_model=HealthResponse)
async def get_system_health(
    context: SecurityContext = Depends(get_security_context)
) -> HealthResponse:
    """Get comprehensive system health status"""
    
    if Permission.TELEMETRY_READ not in context.permissions:
        from fastapi import HTTPException
        raise HTTPException(status_code=403, detail="Permission required: telemetry:read")
    
    current_time = datetime.utcnow()
    overall_uptime = int(time.time() - system_start_time)
    
    # Update component states
    components = {}
    overall_healthy = True
    
    for component_name, state in component_states.items():
        # Simulate component health checks
        component_uptime = overall_uptime
        error_count = state["errors"]
        
        # Determine component status based on simulated conditions
        if error_count > 10:
            status = "unhealthy"
            overall_healthy = False
        elif error_count > 5:
            status = "degraded" 
            overall_healthy = False
        else:
            status = "healthy"
        
        # Generate appropriate status details
        if status == "healthy":
            details = f"{component_name.replace('_', ' ').title()} operating normally"
        elif status == "degraded":
            details = f"{component_name.replace('_', ' ').title()} experiencing minor issues"
        else:
            details = f"{component_name.replace('_', ' ').title()} experiencing critical issues"
        
        components[component_name] = ComponentHealth(
            status=status,
            details=details,
            last_check=current_time,
            uptime_seconds=component_uptime,
            error_count=error_count
        )
        
        # Update state
        state["last_check"] = current_time
        state["status"] = status
        state["uptime"] = component_uptime
    
    # Determine overall system status
    if overall_healthy:
        overall_status = "healthy"
    elif any(c.status == "unhealthy" for c in components.values()):
        overall_status = "unhealthy"
    else:
        overall_status = "degraded"
    
    return HealthResponse(
        status=overall_status,
        components=components,
        timestamp=current_time,
        overall_uptime_seconds=overall_uptime
    )


@router.get("/metrics", response_model=MetricsResponse)
async def get_system_metrics(
    context: SecurityContext = Depends(get_security_context),
    metric_name: Optional[str] = Query(None, description="Filter by specific metric name"),
    time_range_hours: int = Query(1, ge=1, le=168, description="Time range in hours")
) -> MetricsResponse:
    """Get system performance and operational metrics"""
    
    if Permission.TELEMETRY_READ not in context.permissions:
        from fastapi import HTTPException
        raise HTTPException(status_code=403, detail="Permission required: telemetry:read")
    
    current_time = datetime.utcnow()
    
    # Generate system metrics
    metrics = []
    
    # CPU and Memory Metrics
    cpu_metrics = [
        Metric(
            name="cpu_usage_percent",
            value=15.7,
            unit="percent",
            tags={"component": "system", "host": "api-server-1"},
            timestamp=current_time
        ),
        Metric(
            name="memory_usage_percent", 
            value=42.3,
            unit="percent",
            tags={"component": "system", "host": "api-server-1"},
            timestamp=current_time
        ),
        Metric(
            name="memory_usage_bytes",
            value=1073741824,  # 1GB
            unit="bytes",
            tags={"component": "system", "host": "api-server-1"},
            timestamp=current_time
        )
    ]
    
    # API Metrics
    api_metrics = [
        Metric(
            name="http_requests_total",
            value=15847,
            unit="count",
            tags={"method": "GET", "status": "200"},
            timestamp=current_time
        ),
        Metric(
            name="http_request_duration_seconds",
            value=0.089,
            unit="seconds", 
            tags={"method": "POST", "endpoint": "/v1/agents"},
            timestamp=current_time
        ),
        Metric(
            name="http_requests_per_second",
            value=23.7,
            unit="per_second",
            tags={"component": "api_gateway"},
            timestamp=current_time
        )
    ]
    
    # Agent Metrics
    agent_metrics = [
        Metric(
            name="active_agents",
            value=5,
            unit="count",
            tags={"component": "agent_manager"},
            timestamp=current_time
        ),
        Metric(
            name="agent_tasks_completed",
            value=1247,
            unit="count",
            tags={"component": "agent_manager"},
            timestamp=current_time
        ),
        Metric(
            name="agent_success_rate",
            value=0.94,
            unit="ratio",
            tags={"component": "agent_manager"},
            timestamp=current_time
        )
    ]
    
    # Security Metrics
    security_metrics = [
        Metric(
            name="threats_detected_total",
            value=127,
            unit="count",
            tags={"component": "security_engine", "severity": "all"},
            timestamp=current_time
        ),
        Metric(
            name="threats_blocked_total",
            value=98,
            unit="count", 
            tags={"component": "security_engine", "action": "blocked"},
            timestamp=current_time
        ),
        Metric(
            name="security_score",
            value=92.5,
            unit="score",
            tags={"component": "security_engine"},
            timestamp=current_time
        )
    ]
    
    # Intelligence Metrics
    intelligence_metrics = [
        Metric(
            name="ai_decisions_total",
            value=892,
            unit="count",
            tags={"component": "intelligence_brain", "model": "qwen3"},
            timestamp=current_time
        ),
        Metric(
            name="ai_decision_confidence_avg",
            value=0.87,
            unit="ratio",
            tags={"component": "intelligence_brain"},
            timestamp=current_time
        ),
        Metric(
            name="ai_response_time_ms",
            value=142.5,
            unit="milliseconds",
            tags={"component": "intelligence_brain"},
            timestamp=current_time
        )
    ]
    
    # Task Orchestration Metrics
    orchestration_metrics = [
        Metric(
            name="tasks_queued",
            value=12,
            unit="count",
            tags={"component": "orchestrator", "status": "pending"},
            timestamp=current_time
        ),
        Metric(
            name="tasks_completed_total",
            value=2847,
            unit="count",
            tags={"component": "orchestrator", "status": "completed"},
            timestamp=current_time
        ),
        Metric(
            name="task_execution_time_avg_seconds",
            value=187.3,
            unit="seconds",
            tags={"component": "orchestrator"},
            timestamp=current_time
        )
    ]
    
    # Combine all metrics
    all_metrics = (
        cpu_metrics + api_metrics + agent_metrics + 
        security_metrics + intelligence_metrics + orchestration_metrics
    )
    
    # Filter by metric name if requested
    if metric_name:
        all_metrics = [m for m in all_metrics if m.name == metric_name]
    
    return MetricsResponse(
        metrics=all_metrics,
        timestamp=current_time,
        collection_interval_seconds=60
    )


@router.get("/prometheus")
async def get_prometheus_metrics(
    context: SecurityContext = Depends(get_security_context)
) -> str:
    """Get metrics in Prometheus format"""
    
    if Permission.TELEMETRY_READ not in context.permissions:
        from fastapi import HTTPException
        raise HTTPException(status_code=403, detail="Permission required: telemetry:read")
    
    # Generate Prometheus-formatted metrics
    prometheus_metrics = """# HELP xorb_http_requests_total Total HTTP requests
# TYPE xorb_http_requests_total counter
xorb_http_requests_total{method="GET",status="200"} 15847
xorb_http_requests_total{method="POST",status="200"} 3241
xorb_http_requests_total{method="PUT",status="200"} 892
xorb_http_requests_total{method="DELETE",status="200"} 156

# HELP xorb_http_request_duration_seconds HTTP request latency
# TYPE xorb_http_request_duration_seconds histogram
xorb_http_request_duration_seconds_bucket{le="0.1"} 12847
xorb_http_request_duration_seconds_bucket{le="0.5"} 18239
xorb_http_request_duration_seconds_bucket{le="1.0"} 19894
xorb_http_request_duration_seconds_bucket{le="+Inf"} 20136
xorb_http_request_duration_seconds_sum 1792.3
xorb_http_request_duration_seconds_count 20136

# HELP xorb_active_agents Number of active agents
# TYPE xorb_active_agents gauge  
xorb_active_agents 5

# HELP xorb_threats_detected_total Total threats detected
# TYPE xorb_threats_detected_total counter
xorb_threats_detected_total{severity="low"} 45
xorb_threats_detected_total{severity="medium"} 56
xorb_threats_detected_total{severity="high"} 21
xorb_threats_detected_total{severity="critical"} 5

# HELP xorb_security_score Current security score
# TYPE xorb_security_score gauge
xorb_security_score 92.5

# HELP xorb_ai_decisions_total AI decisions made
# TYPE xorb_ai_decisions_total counter
xorb_ai_decisions_total{model="qwen3"} 547
xorb_ai_decisions_total{model="claude"} 345

# HELP xorb_system_uptime_seconds System uptime in seconds
# TYPE xorb_system_uptime_seconds counter
xorb_system_uptime_seconds """ + str(int(time.time() - system_start_time)) + """

# HELP xorb_cpu_usage_percent CPU usage percentage
# TYPE xorb_cpu_usage_percent gauge
xorb_cpu_usage_percent 15.7

# HELP xorb_memory_usage_percent Memory usage percentage  
# TYPE xorb_memory_usage_percent gauge
xorb_memory_usage_percent 42.3
"""
    
    return prometheus_metrics


@router.get("/alerts")
async def get_active_alerts(
    context: SecurityContext = Depends(get_security_context),
    severity: Optional[str] = Query(None, description="Filter by severity"),
    limit: int = Query(50, ge=1, le=1000)
) -> Dict[str, Any]:
    """Get active system alerts"""
    
    if Permission.TELEMETRY_READ not in context.permissions:
        from fastapi import HTTPException
        raise HTTPException(status_code=403, detail="Permission required: telemetry:read")
    
    # Generate sample alerts
    current_time = datetime.utcnow()
    
    alerts = []
    
    # Critical alert example
    if not severity or severity == "critical":
        alerts.append({
            "id": "alert_001",
            "name": "High CPU Usage Detected",
            "severity": "critical",
            "description": "API server CPU usage above 90% for 5 minutes", 
            "component": "api_gateway",
            "timestamp": (current_time - timedelta(minutes=3)).isoformat(),
            "status": "active",
            "tags": {"host": "api-server-1", "threshold": "90%"}
        })
    
    # Warning alerts
    if not severity or severity == "warning":
        alerts.extend([
            {
                "id": "alert_002",
                "name": "Agent Response Time Degraded",
                "severity": "warning", 
                "description": "Average agent response time increased to 2.5 seconds",
                "component": "agent_manager",
                "timestamp": (current_time - timedelta(minutes=15)).isoformat(),
                "status": "active",
                "tags": {"threshold": "2.0s", "current": "2.5s"}
            },
            {
                "id": "alert_003", 
                "name": "Security Threats Increasing",
                "severity": "warning",
                "description": "Threat detection rate 20% above baseline",
                "component": "security_engine",
                "timestamp": (current_time - timedelta(hours=1)).isoformat(), 
                "status": "acknowledged",
                "tags": {"baseline": "12/hour", "current": "14.4/hour"}
            }
        ])
    
    # Apply limit
    alerts = alerts[:limit]
    
    return {
        "alerts": alerts,
        "total": len(alerts),
        "active_count": len([a for a in alerts if a["status"] == "active"]),
        "acknowledged_count": len([a for a in alerts if a["status"] == "acknowledged"]),
        "timestamp": current_time.isoformat()
    }


@router.post("/alerts/{alert_id}/acknowledge")
async def acknowledge_alert(
    alert_id: str,
    context: SecurityContext = Depends(require_permission(Permission.TELEMETRY_WRITE))
) -> Dict[str, str]:
    """Acknowledge a system alert"""
    
    # Simulate alert acknowledgment
    current_time = datetime.utcnow()
    
    return {
        "alert_id": alert_id,
        "status": "acknowledged", 
        "acknowledged_by": context.user_id,
        "acknowledged_at": current_time.isoformat(),
        "message": "Alert acknowledged successfully"
    }


# Health check endpoints for different components
@router.get("/health/api")
async def get_api_health(
    context: SecurityContext = Depends(get_security_context)
) -> Dict[str, Any]:
    """Get API gateway specific health"""
    
    return {
        "component": "api_gateway",
        "status": "healthy",
        "uptime_seconds": int(time.time() - system_start_time),
        "requests_per_second": 23.7,
        "average_response_time_ms": 89,
        "active_connections": 47,
        "last_check": datetime.utcnow().isoformat()
    }


@router.get("/health/agents")
async def get_agents_health(
    context: SecurityContext = Depends(get_security_context)
) -> Dict[str, Any]:
    """Get agent manager health"""
    
    return {
        "component": "agent_manager", 
        "status": "healthy",
        "active_agents": 5,
        "total_agents": 7,
        "agent_success_rate": 0.94,
        "average_task_time_seconds": 187.3,
        "last_check": datetime.utcnow().isoformat()
    }


@router.get("/health/security")
async def get_security_health(
    context: SecurityContext = Depends(get_security_context)
) -> Dict[str, Any]:
    """Get security engine health"""
    
    return {
        "component": "security_engine",
        "status": "healthy", 
        "threats_detected_today": 27,
        "threats_blocked_today": 21,
        "security_score": 92.5,
        "detection_engines_active": 4,
        "last_threat_scan": (datetime.utcnow() - timedelta(minutes=5)).isoformat(),
        "last_check": datetime.utcnow().isoformat()
    }


@router.get("/health/intelligence")
async def get_intelligence_health(
    context: SecurityContext = Depends(get_security_context)
) -> Dict[str, Any]:
    """Get AI intelligence brain health"""
    
    return {
        "component": "intelligence_brain",
        "status": "healthy",
        "active_models": 2,
        "decisions_today": 147,
        "average_confidence": 0.87,
        "average_response_time_ms": 142.5,
        "learning_iterations_today": 23,
        "last_training": (datetime.utcnow() - timedelta(hours=6)).isoformat(),
        "last_check": datetime.utcnow().isoformat()
    }


# Initialize component health states
async def _update_component_health():
    """Periodically update component health states"""
    while True:
        current_time = datetime.utcnow()
        
        # Simulate health state changes
        for component_name, state in component_states.items():
            # Small chance of degradation
            if time.time() % 300 < 1:  # Every 5 minutes roughly
                import random
                if random.random() < 0.1:  # 10% chance
                    state["errors"] += 1
                    
            # Auto-recovery
            if state["errors"] > 0 and time.time() % 120 < 1:  # Every 2 minutes
                import random
                if random.random() < 0.3:  # 30% chance of recovery
                    state["errors"] = max(0, state["errors"] - 1)
            
            state["last_check"] = current_time
            
        await asyncio.sleep(60)  # Update every minute